feat: Make SqlQueryReactor and SqlQueryBase64Reactor siblings extending AbstractSqlQueryReactor
